requirements.txt
torch>=2.0
numpy
pyyaml
soundfile
wandb
awscli
